templates/python-starter/requirements.txt
black
pytest
pre-commit
python-dotenv